pyproject.toml
[project]
name = "sb3-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "pydantic-settings",

    # LangChain / AI
    "langchain",
    "langchain-aws",
    "langchain-community",
    "langchain-core",
    "langchain-postgres",
    "langgraph",
    "strands-agents",

    # Database
    "sqlalchemy",
    "pgvector",
    "redshift-connector",
    "asyncpg",

    # AWS
    "boto3",

    # Auth / HTTP
    "pyjwt",
    "requests",
    "httpx",
    "cryptography",
    "pyopenssl",

    # Misc
    "pyyaml",
    "rich",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sb3_api"]